Make default test session its own target
We cannot compile inside ekat_catch_main.cpp, since we want
to build CatchMain library once for all tests, but we need
to be able to cater all tests created by EkatCreateUnitTest,
regardless of whether the option USER_DEFINED_TEST_SESSION
was used or not.
