+++ cmake/EkatCreateUnitTest.cmake
@@ -99,8 +99,8 @@ function(EkatCreateUnitTestExec exec_name exec_srcs)
   if (NOT ecute_EXCLUDE_MAIN_CPP)
     target_link_libraries(${target_name} PUBLIC ekat::CatchMain)
   endif ()
-  if (ecute_USER_DEFINED_TEST_SESSION)
-    target_compile_definitions(${target_name} PRIVATE USER_DEFINED_TEST_SESSION)
+  if (NOT ecute_USER_DEFINED_TEST_SESSION)
+    target_link_libraries(${target_name} PUBLIC ekat::DefaultTestSession)
   endif ()
   if (ecute_LIBS)
     target_link_libraries(${target_name} PUBLIC "${ecute_LIBS}")
