+++ src/testing-support/ekat_catch_main.cpp
@@ -12,12 +12,9 @@
 #include <mpi.h>
 #endif
 
-#if !defined(USER_DEFINED_TEST_SESSION) && defined(EKAT_HAS_KOKKOS)
-#include "ekat_kokkos_session.hpp"
-#endif
-
-// If USER_DEFINED_TEST_SESSION is defined, the user MUST provide implementations
-// for the following two routines
+// These two functions MUST be defined somewhere. EKAT provides the ekat::TestSession library,
+// which implements them. If users want to define their own, they can implement their library,
+// and PREVENT linking of ekat::TestSession mini-lib
 void ekat_initialize_test_session (int argc, char** argv, const bool print_config);
 void ekat_finalize_test_session ();
 
@@ -101,28 +98,3 @@ int main (int argc, char **argv) {
   // Return test result
   return num_failed != 0 ? 1 : 0;
 }
-
-#ifndef USER_DEFINED_TEST_SESSION
-void ekat_initialize_test_session (int argc, char** argv, const bool print_config) {
-#ifdef EKAT_HAS_KOKKOS
-  ekat::initialize_kokkos_session (argc,argv,print_config);
-#endif
-
-#ifdef EKAT_ENABLE_FPE_DEFAULT_MASK
-  int mask = FE_DIVBYZERO | FE_INVALID | FE_OVERFLOW;
-  ekat::enable_fpes(mask);
-#endif
-
-  if (print_config) {
-    printf(" active avx set: %s\n", ekat::active_avx_string().c_str());
-    printf(" compiler id: %s\n", ekat::compiler_id_string().c_str());
-    printf(" current FPE: mask=%d, %s\n", ekat::get_enabled_fpes(), ekat::fpe_config_string().c_str());
-  }
-}
-
-void ekat_finalize_test_session () {
-#ifdef EKAT_HAS_KOKKOS
-  ekat::finalize_kokkos_session ();
-#endif
-}
-#endif // !defined(USER_DEFINED_TEST_SESSION)
